// trabajador-frontend/src/app/models/trabajador.ts
export interface Trabajador {
    idTrabajador: number;
    nombre: string;
    apellidos: string;
    genero: string;
    fechaNac: string;
    numeroDocumento: string;
    direccion: string;
    telefono: string;
    email: string;
    acceso: string;
    usuario: string;
    clave: string;
  }
  